Plot Cover Extent

Prints a graph of cover extent.

plot_cover(csv, title = "cover_change.png",
  heading = "% Cover Change", leg_pos = "topleft",
  y_lab = "Proportion of Samples at or below X% Cover")

Arguments

csv
is a csv with cols (id, year, type, results(1:50))

title
is the character vector to name the output file (png)

heading
is the character vector used as the main title of the plot

leg_pos
is the legend position argument (character)

y_lab
is the label for the y axis (character)

Value

prints a graph of mean cover extent into the working directory